requirements.txt
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cognite-sdk==7.54.13
cryptography==43.0.0
idna==3.7
msal==1.30.0
numpy==2.0.1
oauthlib==3.2.2
pandas==2.2.2
protobuf==5.27.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
